modify train.py in trainer to run as model pipeline
