+++ requirements.txt
@@ -38,3 +38,7 @@ plotly==5.18.0
 tabulate==0.9.0
 yellowbrick==1.5
 collection==0.1.6
+google-cloud-storage
+gcsfs
+python-dotenv
+dotenv
